VulnViz/requirements.txt
art==5.2
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
fire==0.3.1
Flask==2.0.1
idna==2.10
ifaddr==0.1.7
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
netifaces==0.11.0
py==1.10.0
pyshark==0.4.3
python-libnmap==0.7.2
requests==2.25.0
six==1.16.0
soupsieve==2.2.1
termcolor==1.1.0
urllib3==1.26.2
Werkzeug==2.0.1
